+++ Day_7/library_4.py
@@ -1,0 +1,24 @@
+# Explore the world of third-party packages
+import matplotlib
+print("Matplotlib version:", matplotlib.__version__)
+
+# Package Installes: A package installer is a tool which helps you download,
+# install and manage third-party libraries
+
+# Package Installer: A tool that helps you download, install and manage third-party
+# libraries or packages.
+
+# PIP -> most python installations
+# Pypy -> online python packages
+
+# Using PIP is important and works directly online from the terminal
+
+# Requests -> HTTp requests and is usefly
+# Numpy -> numerical computations -> mathematical operations
+# Pandas -> Data manipulations and
+# matplotlib -> used for data visualisations ( static, interactive plots)
+# beuatifoulsoup -> web-scraping library
+# sqlalchemy -> makes working with SQL databases -> uses python code instead of raw SQL
+# sci-kit learn -> mining and machine
+# pytest -> framework for python code
+# virtualenv -> isolated python environment